move setup_cell_id to another file, and add logic to detect setup cell
