+++ frontend/src/components/editor/actions/useCellActionButton.tsx
@@ -39,12 +39,8 @@ import { Switch } from "@/components/ui/switch";
 import { toast } from "@/components/ui/use-toast";
 import { aiCompletionCellAtom } from "@/core/ai/state";
 import { maybeAddMarimoImport } from "@/core/cells/add-missing-import";
-import {
-  hasOnlyOneCellAtom,
-  SETUP_CELL_ID,
-  useCellActions,
-} from "@/core/cells/cells";
-import type { CellId } from "@/core/cells/ids";
+import { hasOnlyOneCellAtom, useCellActions } from "@/core/cells/cells";
+import { type CellId, SETUP_CELL_ID } from "@/core/cells/ids";
 import type { CellData } from "@/core/cells/types";
 import { formatEditorViews } from "@/core/codemirror/format";
 import { toggleToLanguage } from "@/core/codemirror/language/commands";
